tests/wpt/tests/css/filter-effects/parsing/flood-color-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Filter Effects Module Level 1: parsing flood-color with invalid values</title>
<link rel="help" href="https://drafts.fxtf.org/filter-effects/#FloodColorProperty">
<meta name="assert" content="flood-color supports only the grammar '<color>'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("flood-color", "none");
test_invalid_value("flood-color", "black white");
</script>
</body>
</html>
